backend/requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
sentence-transformers>=2.0.0
faiss-cpu>=1.7.0
numpy>=1.21.0,<2.0.0
pydantic>=1.8.0
google-generativeai>=0.3.0
python-dotenv>=0.19.0
websockets>=10.0
python-multipart>=0.0.5
PyJWT>=2.8.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
boto3>=1.26.0
